fix DX compile

+++ tests/source/RTTShadowScene.cpp
@@ -37,7 +37,7 @@ void RTTSample::init(Widget* parentWidget)
     texNode2->runAction(Repeat::createForever(RotateBy3D::create(Vector3(0.0f, -DEGREE_TO_RADIANS(360.0f), 0.0f), 5.0f)));
     
     // create RTT, move camera to head
-    int rw = screenSize.x / 5, rh = screenSize.y / 5;
+    int rw = int(screenSize.x / 5), rh = int(screenSize.y / 5);
     TexturePtr rttT = ResourceManager::getSingleton().createRenderTexture(rw, rh);
     rttT->setBeforeRenderingFunc([]{
         auto activeCamera = SceneManager::getSingleton().getActiveCamera();
